separate RecordDisposer.cs and VRMEditorExporter.cs
