+++ Assets/VRM/UniVRM/Editor/Format/RecordDisposer.cs
@@ -1,0 +1,147 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using UnityEditor;
+using UnityEngine;
+
+namespace VRM
+{
+    public struct RecordDisposer : IDisposable
+    {
+        int _group;
+        public RecordDisposer(UnityEngine.Object[] objects, string msg)
+        {
+            Undo.IncrementCurrentGroup();
+            _group = Undo.GetCurrentGroup();
+            Undo.RecordObjects(objects, msg);
+        }
+
+        public void Dispose()
+        {
+            Undo.RevertAllDownToGroup(_group);
+        }
+
+        /// <summary>
+        /// VRMを構成するコンポーネントをコピーする。
+        /// </summary>
+        /// <param name="go">コピー元</param>
+        /// <param name="root">コピー先</param>
+        /// <param name="map">コピー元とコピー先の対応関係</param>
+        public static void CopyVRMComponents(GameObject go, GameObject root,
+            Dictionary<Transform, Transform> map)
+        {
+            {
+                // blendshape
+                var src = go.GetComponent<VRMBlendShapeProxy>();
+                if (src != null)
+                {
+                    var dst = root.AddComponent<VRMBlendShapeProxy>();
+                    dst.BlendShapeAvatar = src.BlendShapeAvatar;
+                }
+            }
+
+            {
+                var secondary = go.transform.Find("secondary");
+                if (secondary == null)
+                {
+                    secondary = go.transform;
+                }
+
+                var dstSecondary = root.transform.Find("secondary");
+                if (dstSecondary == null)
+                {
+                    dstSecondary = new GameObject("secondary").transform;
+                    dstSecondary.SetParent(root.transform, false);
+                }
+
+                // 揺れモノ
+                foreach (var src in go.transform.GetComponentsInChildren<VRMSpringBoneColliderGroup>())
+                {
+                    var dst = map[src.transform];
+                    var dstColliderGroup = dst.gameObject.AddComponent<VRMSpringBoneColliderGroup>();
+                    dstColliderGroup.Colliders = src.Colliders.Select(y =>
+                    {
+                        var offset = dst.worldToLocalMatrix.MultiplyPoint(src.transform.localToWorldMatrix.MultiplyPoint(y.Offset));
+                        return new VRMSpringBoneColliderGroup.SphereCollider
+                        {
+                            Offset = offset,
+                            Radius = y.Radius
+                        };
+                    }).ToArray();
+                }
+
+                foreach (var src in go.transform.GetComponentsInChildren<VRMSpringBone>())
+                {
+                    // Copy VRMSpringBone
+                    var dst = dstSecondary.gameObject.AddComponent<VRMSpringBone>();
+                    dst.m_comment = src.m_comment;
+                    dst.m_stiffnessForce = src.m_stiffnessForce;
+                    dst.m_gravityPower = src.m_gravityPower;
+                    dst.m_gravityDir = src.m_gravityDir;
+                    dst.m_dragForce = src.m_dragForce;
+                    if (src.m_center != null)
+                    {
+                        dst.m_center = map[src.m_center];
+                    }
+
+                    dst.RootBones = src.RootBones.Select(x => map[x]).ToList();
+                    dst.m_hitRadius = src.m_hitRadius;
+                    if (src.ColliderGroups != null)
+                    {
+                        dst.ColliderGroups = src.ColliderGroups
+                            .Select(x => map[x.transform].GetComponent<VRMSpringBoneColliderGroup>()).ToArray();
+                    }
+                }
+            }
+
+#pragma warning disable 0618
+            {
+                // meta(obsolete)
+                var src = go.GetComponent<VRMMetaInformation>();
+                if (src != null)
+                {
+                    src.CopyTo(root);
+                }
+            }
+#pragma warning restore 0618
+
+            {
+                // meta
+                var src = go.GetComponent<VRMMeta>();
+                if (src != null)
+                {
+                    var dst = root.AddComponent<VRMMeta>();
+                    dst.Meta = src.Meta;
+                }
+            }
+
+            {
+                // firstPerson
+                var src = go.GetComponent<VRMFirstPerson>();
+                if (src != null)
+                {
+                    src.CopyTo(root, map);
+                }
+            }
+
+            {
+                // humanoid
+                var dst = root.AddComponent<VRMHumanoidDescription>();
+                var src = go.GetComponent<VRMHumanoidDescription>();
+                if (src != null)
+                {
+                    dst.Avatar = src.Avatar;
+                    dst.Description = src.Description;
+                }
+                else
+                {
+                    var animator = go.GetComponent<Animator>();
+                    if (animator != null)
+                    {
+                        dst.Avatar = animator.avatar;
+                    }
+                }
+            }
+        }
+    }
+}
