+++ Assets/VRM/UniVRM/Editor/Format/RecordDisposer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b304ed2aeece5a54191a5a8b69d9d113
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
