+++ Assets/VRM/UniVRM/Editor/Format/VRMHumanoidNormalizerMenu.cs
@@ -49,10 +49,10 @@ namespace VRM
             var go = Selection.activeObject as GameObject;
 
             GameObject normalizedRoot = null;
-            using (new VRMExportSettings.RecordDisposer(go.transform.Traverse().ToArray(), "before normalize"))
+            using (new RecordDisposer(go.transform.Traverse().ToArray(), "before normalize"))
             {
                 var normalized = BoneNormalizer.Execute(go, true, false);
-                VRMExportSettings.CopyVRMComponents(go, normalized.Root, normalized.BoneMap);
+                RecordDisposer.CopyVRMComponents(go, normalized.Root, normalized.BoneMap);
                 normalizedRoot = normalized.Root;
             }
             Selection.activeGameObject = normalizedRoot;
